refactor(request_service): extract NavigationGuards from PuppetCommander
Move intercept wiring, deferred navigation errors, main-frame filtering,
resource abort/continue, and redirect-chain validation into
NavigationGuards. Commander keeps page lifecycle and preload; Policy
rules stay single-homed via ctx.policy.validate_*! (Strong S3).
