+++ lib/html2rss/request_service/puppet_commander.rb
@@ -23,6 +23,7 @@ module Html2rss
         @browser = browser
         @skip_request_resources = skip_request_resources
         @referer = referer
+        @navigation_guards = NavigationGuards.new(ctx:, skip_request_resources:)
       end
 
       ##
@@ -33,9 +34,9 @@ module Html2rss
         page = new_page
         navigation_response = navigate_to_destination(page, ctx.url)
         perform_preload(page)
-        raise_navigation_error_if_any
-        final_navigation_response = latest_navigation_response || navigation_response
-        validate_navigation_response!(final_navigation_response)
+        navigation_guards.raise_deferred_error!
+        final_navigation_response = navigation_guards.latest_navigation_response || navigation_response
+        navigation_guards.validate_final!(final_navigation_response)
         build_response(page, final_navigation_response)
       ensure
         page&.close
@@ -46,9 +47,8 @@ module Html2rss
       # @see https://yusukeiwaki.github.io/puppeteer-ruby-docs/Puppeteer/Page.html
       def new_page
         page = browser.new_page
-        @main_frame = page.main_frame if page.respond_to?(:main_frame)
         configure_page(page)
-        configure_navigation_guards(page)
+        navigation_guards.install!(page)
         page
       end
 
@@ -62,28 +62,16 @@ module Html2rss
       end
 
       ##
-      # @param page [Puppeteer::Page]
-      # @return [void]
-      def configure_navigation_guards(page)
-        page.request_interception = true
-        page.on('request') do |request|
-          handle_request(request)
-        end
-        page.on('response') { |response| handle_response(response) }
-      end
-
-      ##
       # @param page [Puppeteer::Page] browser page
       # @param url [Html2rss::Url] target URL
       # @return [Puppeteer::HTTPResponse, nil] the navigation response if one was produced
       def navigate_to_destination(page, url)
-        @navigation_error = nil
-        @latest_navigation_response = nil
+        navigation_guards.begin_navigation!
         page.goto(url, wait_until: 'networkidle0', referer:, timeout: navigation_timeout_ms).tap do
-          raise_navigation_error_if_any
+          navigation_guards.raise_deferred_error!
         end
       rescue StandardError
-        raise_navigation_error_if_any
+        navigation_guards.raise_deferred_error!
 
         raise
       end
@@ -95,15 +83,7 @@ module Html2rss
 
       private
 
-      attr_reader :ctx, :browser, :skip_request_resources, :referer, :latest_navigation_response, :main_frame
-
-      ##
-      # Re-raises a deferred navigation error when one was captured.
-      #
-      # @raise [Html2rss::Error] when a navigation request or response validation failed
-      def raise_navigation_error_if_any
-        raise @navigation_error if @navigation_error
-      end
+      attr_reader :ctx, :browser, :skip_request_resources, :referer, :navigation_guards
 
       def navigation_timeout_ms
         timeout = ctx.budget.remaining_timeout_seconds || ctx.policy.total_timeout_seconds
@@ -116,40 +96,6 @@ module Html2rss
         ctx.headers.reject { |key, _| BROWSER_UNSAFE_HEADERS.include?(key.to_s.downcase) }
       end
 
-      def handle_request(request)
-        validate_request!(request)
-
-        skip_request_resources.member?(request.resource_type) ? request.abort : request.continue
-      rescue Html2rss::Error => error
-        store_navigation_error(error, navigation_request: request.navigation_request?)
-        request.abort
-      end
-
-      def handle_response(response)
-        @latest_navigation_response = response if main_frame_navigation_response?(response)
-        validate_response!(response)
-      rescue Html2rss::Error => error
-        store_navigation_error(error, navigation_request: response.request.navigation_request?)
-      end
-
-      def validate_request!(request)
-        validate_navigation_redirect_chain!(request)
-        validate_navigation_target!(request)
-      end
-
-      def main_frame_navigation_response?(response)
-        request = response.request
-        return false unless request.navigation_request?
-        return true unless request.respond_to?(:frame)
-
-        frame = request.frame
-        return true if frame.nil?
-        return frame == main_frame unless main_frame.nil?
-        return true unless frame.respond_to?(:parent_frame)
-
-        frame.parent_frame.nil?
-      end
-
       def build_response(page, navigation_response)
         Response.new(
           body: body(page),
@@ -159,46 +105,9 @@ module Html2rss
         )
       end
 
-      def validate_navigation_response!(navigation_response)
-        final_url = response_url(navigation_response, ctx.url)
-        ctx.policy.validate_remote_ip!(ip: remote_ip(navigation_response), url: final_url)
-      end
-
-      def validate_response!(response)
-        validate_navigation_response!(response)
-      end
-
       def response_url(navigation_response, fallback_url)
         raw_url = navigation_response&.url || fallback_url.to_s
         Html2rss::Url.from_absolute(raw_url)
-      end
-
-      def remote_ip(navigation_response)
-        navigation_response.remote_address&.ip
-      end
-
-      def request_chain(request)
-        (request.redirect_chain + [request]).map { |entry| request_url(entry) }
-      end
-
-      def request_url(request)
-        Html2rss::Url.from_absolute(request.url)
-      end
-
-      def validate_navigation_redirect_chain!(request)
-        request_chain(request).each_cons(2) do |from_url, to_url|
-          ctx.policy.validate_redirect!(from_url:, to_url:, origin_url: ctx.origin_url, relation: ctx.relation)
-        end
-      end
-
-      def validate_navigation_target!(request)
-        ctx.policy.validate_request!(url: request_url(request), origin_url: ctx.origin_url, relation: ctx.relation)
-      end
-
-      def store_navigation_error(error, navigation_request:)
-        return unless navigation_request
-
-        @navigation_error = error if @navigation_error.nil?
       end
 
       def perform_preload(page)
